remove preloading images
