+++ app/src/main/java/de/westnordost/streetcomplete/data/Preloader.kt
@@ -2,7 +2,6 @@ package de.westnordost.streetcomplete.data
 
 import de.westnordost.countryboundaries.CountryBoundaries
 import de.westnordost.osmfeatures.FeatureDictionary
-import de.westnordost.streetcomplete.screens.main.map.MapIcons
 import de.westnordost.streetcomplete.util.ktx.format
 import de.westnordost.streetcomplete.util.ktx.nowAsEpochMilliseconds
 import de.westnordost.streetcomplete.util.logs.Log
@@ -15,7 +14,6 @@ import kotlinx.coroutines.withContext
 class Preloader(
     private val countryBoundaries: Lazy<CountryBoundaries>,
     private val featuresDictionary: Lazy<FeatureDictionary>,
-    private val mapIcons: MapIcons,
 ) {
 
     suspend fun preload() {
@@ -26,8 +24,6 @@ class Preloader(
             // names dictionary is necessary when displaying an element that has no name or
             // when downloading the place name quest (etc)
             launch { preloadFeatureDictionary() }
-            // map icons should be displayed as soon as the map finished loading its style
-            launch { createMapIcons() }
         }
 
         Log.i(TAG, "Preloading data took ${((nowAsEpochMilliseconds() - time) / 1000.0).format(1)}s")
@@ -45,15 +41,6 @@ class Preloader(
         countryBoundaries.value
         val seconds = (nowAsEpochMilliseconds() - time) / 1000.0
         Log.i(TAG, "Loaded country boundaries in ${seconds.format(1)}s")
-    }
-
-    private suspend fun createMapIcons() = withContext(Dispatchers.Default) {
-        val time = nowAsEpochMilliseconds()
-        mapIcons.pinBitmaps
-        mapIcons.presetBitmaps
-        mapIcons.markerBitmaps
-        val seconds = (nowAsEpochMilliseconds() - time) / 1000.0
-        Log.i(TAG, "Created map icons in ${seconds.format(1)}s")
     }
 
     companion object {
