+++ app/src/main/java/de/westnordost/streetcomplete/screens/main/MainModule.kt
@@ -3,7 +3,6 @@ package de.westnordost.streetcomplete.screens.main
 import de.westnordost.streetcomplete.data.location.RecentLocationStore
 import de.westnordost.streetcomplete.screens.main.edithistory.EditHistoryViewModel
 import de.westnordost.streetcomplete.screens.main.edithistory.EditHistoryViewModelImpl
-import de.westnordost.streetcomplete.screens.main.map.MapIcons
 import de.westnordost.streetcomplete.util.location.LocationAvailabilityReceiver
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.core.qualifier.named
@@ -12,8 +11,6 @@ import org.koin.dsl.module
 val mainModule = module {
     single { LocationAvailabilityReceiver(get()) }
     single { RecentLocationStore() }
-
-    single { MapIcons(get(), get()) }
 
     viewModel<MainViewModel> { MainViewModelImpl(
         get(), get(), get(), get(), get(), get(), get(), get(), get(), get(), get(), get(), get()
